requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
astroid==3.2.4
azure-core==1.31.0
azure-storage-blob==12.23.1
# backports.zoneinfo==0.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
black==24.8.0
boto3==1.35.44
botocore==1.35.44
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
condition==1.0.8
cryptography==43.0.1
defusedxml==0.8.0rc2
dill==0.3.8
Django==4.2.15
django-allauth==64.2.0
django-filter==24.3
django-rest-framework-mongoengine==3.4.1
django-social-auth==0.7.28
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
ecdsa==0.19.0
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.8
isodate==0.7.2
isort==5.13.2
jmespath==1.0.1
mccabe==0.7.0
mongoengine==0.29.0
mypy-extensions==1.0.0
# numpy==1.24.4
oauth2==1.9.0.post1
oauthlib==3.2.2
packaging==24.1
# pandas==2.0.3
pathspec==0.12.1
phonenumbers==8.13.44
pillow==10.4.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pylint==3.2.7
pymongo==4.8.0
pyotp==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-openid==2.2.5
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.3
six==1.16.0
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
versioneer==0.29
